task2/resolve/requirements.txt
opencv-python==4.9.0.80
Pillow==10.4.0
numpy==1.26.4

